requirements.txt
piexif==1.1.3
Pillow==11.1.0
PyQt5==5.15.11
PyQt5_sip==12.15.0
pytest==8.3.4
python-dotenv==1.0.1
Requests==2.32.3

ipython~=8.12.3
traitlets~=5.14.3
openai~=1.50.0